==> USAuthors/3_Analysis for DisCA/0.1_first pass and adjustment.rmd <==
---
title: "Authors data - CA"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ExPosition)
library(PTCA4CATA)
library(ggplot2)
library(ggrepel)
library(tidyverse)
library(plotly)
```

## Data

```{r}
load("Data/AllPunctCount.sDisCA.rda")

knitr::kable(matrix(rownames(AllPunctCount.sDisCA), ncol = 7))
```

### English

```{r}
knitr::kable(AllPunctCount.sDisCA.en)
```

### French

```{r}
knitr::kable(AllPunctCount.sDisCA.fr)
```

### Combined

```{r}
knitr::kable(AllPunctCount.sDisCA)
```

## CA

### data cleaning
```{r}
alldat <- AllPunctCount.sDisCA[rowSums(AllPunctCount.sDisCA) > 0,]
knitr::kable(AllPunctCount.sDisCA[rowSums(AllPunctCount.sDisCA) == 0,])

knitr::kable(alldat[duplicated(alldat) == TRUE,])
alldat <- alldat[duplicated(alldat) == FALSE,]

## combine - and n-dash
alldat.78 <- cbind(alldat[, c(1,2,3,4,5,6,9,10)], alldat[,7] + alldat[,8])
colnames(alldat.78)[9] <- "--"

```

### epCA
```{r}
ca.res <- epCA(alldat.78, symmetric = TRUE, graphs = FALSE)
```

#### Plot results
```{r}
ca.plot <- createFactorMapIJ(ca.res$ExPosition.Data$fi, ca.res$ExPosition.Data$fj,axis1 = 2, axis2 = 3,
                            text.cex.i = 3,
                            col.points.i = "maroon",
                            col.labels.i = "maroon",
                            title = "Picked Authors - Symmetric")

ca.label <- createxyLabels.gen(2,3,
                               lambda = ca.res$ExPosition.Data$eigs,
                               tau = round(ca.res$ExPosition.Data$t),
                               axisName = "Component "
)

ca.plot$baseMap + ca.label + ca.plot$I_points + ca.plot$J_points + ca.plot$J_labels

```
#### Authors
```{r}
ca.plot <- createFactorMap(ca.res$ExPosition.Data$fi,axis1 = 2, axis2 = 3,
                           text.cex = 3,
                           col.points = "maroon",
                           col.labels = "maroon",
                           title = "Picked Authors - Symmetric")#,

ca.plot$zeMap + ca.label

## try interactive plot
fi2plot <- ca.res$ExPosition.Data$fi[,1:2] %>% as.data.frame
fj2plot <- ca.res$ExPosition.Data$fj[,1:2] %>% as.data.frame
colnames(fj2plot) <- colnames(fi2plot) <- paste0("Component", c(1,2))
p <- ggplot(data = fi2plot, aes(x = Component1, y = Component2, text = rownames(ca.res$ExPosition.Data$fi))) +
  geom_point(color = "maroon") + theme_bw()
q <- ggplot(fj2plot, aes(x = Component1, y = Component2, label = rownames(ca.res$ExPosition.Data$fj))) +
  geom_point(color = "darkolivegreen", pch = 18, size = 3) + 
  geom_text_repel(size = 7) + theme_bw()

p2 <- ggplot() +
  geom_point(data = fi2plot, aes(x = Component1, y = Component2, label = rownames(ca.res$ExPosition.Data$fi)), color = "maroon") +
  geom_point(data = fj2plot, aes(x = Component1, y = Component2, label = rownames(ca.res$ExPosition.Data$fj)), color = "darkolivegreen", pch = 18, size = 3) + 
  geom_text_repel(data = fj2plot, aes(x = Component1, y = Component2, label = rownames(ca.res$ExPosition.Data$fj)), color = "darkolivegreen", size = 7) + theme_bw()

ggplotly(p, tooltip = c("label"))
```

#### Punctuations
```{r}
ca.j.plot <- createFactorMap(ca.res$ExPosition.Data$fj,
                             axis1 = 2, axis2 = 3, 
                             text.cex = 10,
                             col.points = "darkolivegreen",
                             col.labels = "darkolivegreen",
                             pch = 18,
                             title = "Picked Authors - Symmetric")#,

ca.j.plot$zeMap + ca.label

```


```{r}
elimdat.78 <- alldat.78[rowsum4hist > 2000,]
author.keep <- rownames(elimdat.78)

keep.en <- AllPunctCount.sDisCA.en[which(rownames(AllPunctCount.sDisCA.en) %in% author.keep),]
keep.en <- keep.en[duplicated(keep.en) == FALSE,]

keep.fr <- AllPunctCount.sDisCA.fr[which(rownames(AllPunctCount.sDisCA.fr) %in% author.keep),]
keep.fr <- keep.fr[duplicated(keep.fr) == FALSE,]

write.csv(keep.en, file = "Data/keepEnAuthor.csv")
write.csv(keep.fr, file = "Data/keepFrAuthor.csv")
```
